+++ rdf_differ/loader/services/loader.py
@@ -15,13 +15,13 @@ from rdf_differ import config
 from rdf_differ.core.domain.constants import DeltaOp, mime_type_for
 from rdf_differ.loader.adapters import sparql_queries as q
 from rdf_differ.loader.adapters.graph_store import GraphStorePort
+from rdf_differ.loader.domain.exceptions import ValidationError
 from rdf_differ.loader.domain.model import (
     BlankNodePolicy,
     BlankNodeStrategy,
     DeltaCounts,
     IdentityBlankNodeStrategy,
     LoadResult,
-    ValidationError,
     VersionSpec,
     VersionStoreConfig,
     consecutive_pairs,